requirements.txt
Matplotlib
Pillow
Keras
torch
torchvision
torchaudio
tensorflow==2.9.1
requests
numpy
opencv-python
visualkeras
pydot
eli5
seaborn
sklearn
keract
graphviz
keras-visualizer

